Add a shared CdaParamsTable for hook docs
